requirements.txt
django 
djangorestframework 
django-cors-headers
psycopg2
psycopg2-binary
bs4